Moving shared utils and renaming Fixes #94
