+++ src/shared/utils/index.ts
@@ -1,5 +1,3 @@
-import { Period, PeriodFrequency } from 'background/service/networkModel';
-
 import BroadcastChannelMessage from './message/broadcastChannelMessage';
 import PortMessage from './message/portMessage';
 
@@ -19,30 +17,3 @@ const format = (str, ...args) => {
 };
 
 export { Message, t, format };
-
-// const chainsDict = keyBy(CHAINS, 'serverId');
-// export const getChain = (chainId?: string) => {
-//   if (!chainId) {
-//     return null;
-//   }
-//   return chainsDict[chainId];
-// };
-
-export const getPeriodFrequency = (period: Period): PeriodFrequency => {
-  switch (period) {
-    case Period.oneDay:
-      return PeriodFrequency.halfHour;
-    case Period.oneWeek:
-      return PeriodFrequency.oneHour;
-    case Period.oneMonth:
-      return PeriodFrequency.oneDay;
-    case Period.threeMonth:
-      return PeriodFrequency.oneDay;
-    case Period.oneYear:
-      return PeriodFrequency.threeDay;
-    case Period.all:
-      return PeriodFrequency.oneWeek;
-    default:
-      return PeriodFrequency.oneDay;
-  }
-};
